Change type Id field in ChatInfo and GroupChatInfo. Change type variable chatId in TelegramBotClient

+++ NetTelebot.Tests/Utils/WindowsCredential.cs
@@ -31,7 +31,7 @@ namespace NetTelebot.Tests.Utils
             Credential credential = new Credential { Target = botAlias };
 
             return credential.Load()
-                ? new TelegramCredentials { Token = credential.Password, ChatId = Convert.ToInt32(credential.Username) }
+                ? new TelegramCredentials { Token = credential.Password, ChatId = Convert.ToInt64(credential.Username) }
                 : new TelegramCredentials { Token = null, ChatId = 0 };
         }
     }
@@ -54,7 +54,7 @@ namespace NetTelebot.Tests.Utils
         /// This is a sample of the response when you add your BOT into a group. 
         /// 4. Use the "id" of the "chat" object to send your messages.
         /// </summary>
-        public int ChatId { get; set; }
+        public long ChatId { get; set; }
 
         public bool Equals(TelegramCredentials credentials)
         {
